refactor: useArticles 훅 추가
